+++ dev/com.ibm.ws.security.mp.jwt.1.1.config/.settings/bndtools.core.prefs
@@ -1,0 +1,2 @@
+compileErrorAction=build
+eclipse.preferences.version=1
