+++ dev/com.ibm.ws.security.mp.jwt.1.1.config/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+encoding/<project>=UTF-8
