+++ dev/com.ibm.ws.security.mp.jwt.1.1.config/bnd.bnd
@@ -1,0 +1,75 @@
+#*******************************************************************************
+# Copyright (c) 2018 IBM Corporation and others.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License v1.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+#
+# Contributors:
+#     IBM Corporation - initial API and implementation
+#*******************************************************************************
+-include= ~../cnf/resources/bnd/bundle.props
+bVersion=1.0
+javac.source: 1.8
+javac.target: 1.8
+
+Bundle-Name: WAS Secuity MP JWT 1.1 config
+Bundle-SymbolicName: com.ibm.ws.security.mp.jwt.1.1.config
+Bundle-Description: WAS security jwt authentication mpConfig proxy for microprofile, version=${bVersion}
+Require-Capability: osgi.ee; filter:="(&(osgi.ee=JavaSE)(version=1.8))"
+
+#Web-ContextPath: /ibm/api/social-login
+
+
+WS-TraceGroup: \
+  MPJWT
+
+# For each exported package, create (in that package) a package-info.java...
+# file, and place an @version javadoc tag in package-level javadoc. 
+# Append ";provide:=true" if this bundle also provides an implementation
+# for the exported API.
+
+
+Export-Package: \
+   com.ibm.ws.security.mp.jwt.v11.config, \
+   com.ibm.ws.security.mp.jwt.v11.config.impl
+
+Import-Package: \
+   com.ibm.ws.security.mp.jwt.v11, \
+   *
+
+Private-Package: \
+    com.ibm.ws.security.mp.jwt.v11.config.resources.*,
+    
+-dsannotations: com.ibm.ws.security.mp.jwt.v11.config.impl.MpConfigProxyServiceImpl
+
+-dsannotations-inherit = true
+
+
+#Include-Resource: \
+#    OSGI-INF=resources/OSGI-INF
+
+-buildpath: \
+	com.ibm.ws.org.osgi.annotation.versioning;version=latest, \
+        com.ibm.websphere.org.osgi.core;version=latest, \
+        com.ibm.websphere.org.osgi.service.cm;version=latest, \
+        com.ibm.websphere.org.osgi.service.component;version=latest, \
+        com.ibm.wsspi.org.osgi.service.component.annotations;version=latest, \
+        com.ibm.ws.logging;version=latest, \
+        com.ibm.ws.kernel.service;version=latest, \
+        com.ibm.ws.resource;version=latest, \
+        com.ibm.ws.security.mp.jwt.1.1;version=latest, \
+        com.ibm.ws.microprofile.config.1.2;version=latest, \
+        com.ibm.websphere.org.eclipse.microprofile.config.1.3;version=latest
+
+-testpath: \
+    ../build.sharedResources/lib/junit/old/junit.jar;version=file, \
+    com.ibm.ws.junit.extensions;version=latest, \
+    org.jmock:jmock-legacy;version=2.5.0, \
+    cglib:cglib-nodep;version=3.2.7, \
+    org.hamcrest:hamcrest-all;version=1.3, \
+    org.jmock:jmock-junit4;strategy=exact;version=2.5.1, \
+    org.jmock:jmock;strategy=exact;version=2.5.1, \
+    com.ibm.ws.org.objenesis:objenesis;version=1.0, \
+    com.ibm.ws.kernel.boot;version=latest, \
+    com.ibm.ws.org.jose4j.0.5.1;version=latest
