+++ dev/com.ibm.ws.security.mp.jwt.1.1.config/build.gradle
@@ -1,0 +1,6 @@
+plugins {
+  id "net.ossindex.audit" version "0.3.8-beta"
+  }
+audit {
+    failOnError = false
+}
