+++ dev/com.ibm.ws.security.mp.jwt.1.1.config/build.xml
@@ -1,0 +1,13 @@
+<!--
+    Copyright (c) 2018 IBM Corporation and others.
+    All rights reserved. This program and the accompanying materials
+    are made available under the terms of the Eclipse Public License v1.0
+    which accompanies this distribution, and is available at
+    http://www.eclipse.org/legal/epl-v10.html
+   
+    Contributors:
+        IBM Corporation - initial API and implementation
+ -->
+<project name="com.ibm.ws.security.mp.jwt.1.1" default="build">
+	<import file="../ant_build/public_imports/standard_imports.xml"/>
+</project>
