+++ dev/com.ibm.ws.security.mp.jwt.1.1.config/resources/com/ibm/ws/security/mp/jwt/v11/config/resources/MpConfigProxyMessages.nlsprops
@@ -1,0 +1,31 @@
+###############################################################################
+# Copyright (c) 2018 IBM Corporation and others.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License v1.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+#
+# Contributors:
+#     IBM Corporation - initial API and implementation
+###############################################################################
+#CMVCPATHNAME com.ibm.ws.security/resources/com/ibm/ws/security/mp/jwt/v11/resources/MpConfigProxyMessages.nlsprops
+#COMPONENTPREFIX CWWKS
+#COMPONENTNAMEFOR WebSphere Application Server MicroProfile JWT 1.1 Config Proxy
+#ISMESSAGEFILE TRUE
+#NLS_MESSAGEFORMAT_VAR
+#NLS_ENCODING=UNICODE
+# -------------------------------------------------------------------------------------------------
+# Message prefix block: CWWKS5775 - CWWKS5799
+
+MPJWT_11_CONFIG_PROXY_PROCESSED=CWWKS5775I: The MicroProfile JWT v11 mpConfigProxy was successfully processed.
+MPJWT_11_CONFIG_PROXY_PROCESSED.explanation=The indicated configuration was successfully processed.
+MPJWT_11_CONFIG_PROXY_PROCESSED.useraction=No action is required.
+
+MPJWT_11_CONFIG_PROXY_MODIFIED=CWWKS5776I: The MicroProfile JWT v11 mpConfigProxy was successfully processed.
+MPJWT_11_CONFIG_PROXY_MODIFIED.explanation=The indicated configuration was successfully processed.
+MPJWT_11_CONFIG_PROXY_MODIFIED.useraction=No action is required.
+
+MPJWT_11_CONFIG_PROXY_DEACTIVATED=CWWKS5777I: The MicroProfile JWT v11 mpConfigProxy was successfully deactivated.
+MPJWT_11_CONFIG_PROXY_DEACTIVATED.explanation=The indicated configuration was successfully deactivated.
+MPJWT_11_CONFIG_PROXY_DEACTIVATED.useraction=No action is required.
+
