+++ dev/com.ibm.ws.security.mp.jwt/src/com/ibm/ws/security/mp/jwt/MpJwtExtensionService.java
@@ -8,13 +8,23 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.security.mp.jwt.v11;
+package com.ibm.ws.security.mp.jwt;
 
-/**
- *
- */
-public interface TraceConstants {
+import java.util.NoSuchElementException;
+public interface MpJwtExtensionService {
 
-    public final static String TRACE_GROUP = "MPJWT";
-    public final static String MESSAGE_BUNDLE = "com.ibm.ws.security.mp.jwt.v11.resources.MpJwtMessages";
+    /**
+     * @return
+     */
+    public String getVersion();
+
+    /**
+     * @return
+     */
+    public boolean isMpConfigAvailable();
+
+    /**
+     * @return
+     */
+    public <T> T getConfigValue(String propertyName, Class<T> propertyType) throws IllegalArgumentException, NoSuchElementException;
 }
